Add ansible for imaging and ssh config
This commit adds a sample ansible setup, including a playbook, for
imaging an SD card and then enabling ssh. This currently makes the
assumption that a Raspbian-like OS is used.

This setup was tested manually by provisiong an SD card with Raspbian
Stretch Lite and running on a Raspberry Pi 3.