Enable PTUUID Verification
Previously, pi-er users would be required to provide a device which would be
written to. Between different hosts (or in some cases, even on the same hosts,
but with different configurations), the device may be different. As such, this
could result in potentially dangerous operations if a user mistakenly does not
update the device to use.

This change introduces a PTUUID check for GPT devices, allowing a user can
specify a PTUUID alongside their device, and writing operations will fail if
the expected PTUUID is not the PTUUID of the device.