Retry on unexpected PTs
After modifications to the disk have been written, the kernel is
informed of these changes, and it takes a non-zero amount of time for
this information to propagate through. This means that there is a race
been the kernel and pi-er.

This commit ensures that pi-er retries when the kernel is missing
information about the partition table.

This commit has been tested manually and both cases where the race was
evident (and not) have both been observed working with this commit now.